requirements.txt
asgiref==3.3.4
cli-helpers==2.1.0
click==8.0.1
configobj==5.0.6
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
humanize==3.7.1
pgcli==1.9.0
pgspecial==1.11.5
prompt-toolkit==1.0.18
psycopg2-binary==2.8.5
Pygments==2.9.0
pytz==2021.1
setproctitle==1.2.2
six==1.16.0
sqlparse==0.2.4
tabulate==0.8.9
terminaltables==3.1.0
wcwidth==0.2.5
